+++ PIR/include/Module.h
@@ -487,6 +487,7 @@ namespace pcit::pir{
 			core::StepAlloc<Call, uint32_t> calls{};
 			core::StepAlloc<CallVoid, uint32_t> call_voids{};
 			core::StepAlloc<Ret, uint32_t> rets{};
+			core::StepAlloc<CondBranch, uint32_t> cond_branches{};
 			core::StepAlloc<Load, uint32_t> loads{};
 			core::StepAlloc<Store, uint32_t> stores{};
 			core::StepAlloc<CalcPtr, uint32_t> calc_ptrs{};